cleanup frontend
