--- src/cli/mod.rs
@@ -1,8 +1,0 @@
-pub mod node;
-pub use node::*;
-pub mod wallet;
-pub use wallet::*;
-pub mod cardano_command;
-pub use cardano_command::CardanoCommand;
-pub mod start_cli;
-pub use start_cli::Cli;
